\name{PrasathGraph-package}
\alias{PrasathGraph-package}
\alias{PrasathGraph}
\docType{package}
\title{ PrasathGraph }
\description{

This package contains functions to visualize pathway heatmaps and to manipulate pathway 
graphNEL objects, such as identifying upstream and downstream nodes.

}
\details{
\tabular{ll}{
Package: \tab PrasathGraph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-05\cr
License: \tab GPL-2\cr
}

}
\author{
Prasath Pararajalingam

Maintainer: Prasath Pararajalingam <ppararaj@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
